% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getAllTranscriptionFactors,TrenaProject-method}
\alias{getAllTranscriptionFactors,TrenaProject-method}
\alias{getAllTranscriptionFactors}
\title{get all genes annotated by GO to}
\usage{
\S4method{getAllTranscriptionFactors}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
get all genes annotated by GO to
}
