% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getCovariateDataset,TrenaProject-method}
\alias{getCovariateDataset,TrenaProject-method}
\alias{getCovariateDataset}
\title{Get the specified covariates table}
\usage{
\S4method{getCovariateDataset}{TrenaProject}(obj, datasetName)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{datasetName}{character string, the covariate dataset of interest}
}
\description{
Get the specified covariates table
}
