% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getCovariateDatasetNames,TrenaProject-method}
\alias{getCovariateDatasetNames,TrenaProject-method}
\alias{getCovariateDatasetNames}
\title{List the RData files in the covariate directory}
\usage{
\S4method{getCovariateDatasetNames}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
List the RData files in the covariate directory
}
