% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getExpressionDirectory,TrenaProject-method}
\alias{getExpressionDirectory,TrenaProject-method}
\alias{getExpressionDirectory}
\title{Get the names of the expression directory, wherein serialized expression matrices will be found}
\usage{
\S4method{getExpressionDirectory}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
Get the names of the expression directory, wherein serialized expression matrices will be found
}
