% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getFootprintDatabasePort,TrenaProject-method}
\alias{getFootprintDatabasePort,TrenaProject-method}
\alias{getFootprintDatabasePort}
\title{get the port  on which the footprints Postgres database is running}
\usage{
\S4method{getFootprintDatabasePort}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
get the port  on which the footprints Postgres database is running
}
