% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getGenomeName,TrenaProject-method}
\alias{getGenomeName,TrenaProject-method}
\alias{getGenomeName}
\title{get the genome name - to be set by derived classes}
\usage{
\S4method{getGenomeName}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
get the genome name - to be set by derived classes
}
