% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\name{getProjectName,TrenaProject-method}
\alias{getProjectName,TrenaProject-method}
\alias{getProjectName}
\title{get the project name - to be set by derived classes}
\usage{
\S4method{getProjectName}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
get the project name - to be set by derived classes
}
