% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getGenomicRegionsDatasetNames,TrenaProject-method}
\alias{getGenomicRegionsDatasetNames,TrenaProject-method}
\alias{getRegionDatasetNames}
\alias{getGenomicRegionsDataset,TrenaProject-method}
\title{Get a list of datasets each describing regions of interest, typically regulatory regions, open chromatin, etc}
\usage{
\S4method{getGenomicRegionsDatasetNames}{TrenaProject}(obj)

\S4method{getGenomicRegionsDataset}{TrenaProject}(obj, datasetName)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{obj}{An object of class TrenaProject}
}
\description{
Get a list of datasets each describing regions of interest, typically regulatory regions, open chromatin, etc

Get a list of datasets each describing regions of interest, typically regulatory regions, open chromatin, etc
}
