% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getTranscriptsTable,TrenaProject-method}
\alias{getTranscriptsTable,TrenaProject-method}
\alias{getTranscriptsTable}
\title{Get the transcript info for the gene, just the first row}
\usage{
\S4method{getTranscriptsTable}{TrenaProject}(obj,
  targetGene = NA_character_)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
Get the transcript info for the gene, just the first row
}
