% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\name{show,TrenaProject-method}
\alias{show,TrenaProject-method}
\alias{show}
\title{get a summary of the objects}
\usage{
\S4method{show}{TrenaProject}(object)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
get a summary of the objects
}
